More code reorganisation. Started on refactoring table front end code.
