+++ multiqc/templates/default/general_stats.html
@@ -8,96 +8,6 @@ This block generates the General Statistics table at the top of the report.
 
 <div id="general_stats">
   <h2>General Statistics</h2>
-  Table:
   {{ report.general_stats_html }}
 </div>
 
-{% if report.general_stats_beeswarm_html %}
-<div id="general_stats">
-  <h2>General Statistics</h2>
-  {{ report.general_stats_beeswarm_html }}
-</div>
-{% endif %}
-
-
-{% if report.general_stats_html['headers']|length > 0 %}
-<div id="general_stats">
-<h2>
-  General Statistics &nbsp;
-  <button type="button" id="general_stats_copy_btn" class="btn btn-default btn-sm" data-clipboard-target="#general_stats_table"><span class="glyphicon glyphicon-copy"></span> Copy table</button>
-  <button type="button" id="general_stats_colsort_btn" class="btn btn-default btn-sm" data-toggle="modal" data-target="#general_stats_colsort_modal"><span class="glyphicon glyphicon-th"></span> Configure Columns</button>
-  <button type="button" class="btn btn-default btn-sm" id="mqc_genstat_sort_highlight" data-direction="desc" style="display:none;"><span class="glyphicon glyphicon-sort-by-attributes-alt"></span> Sort by highlight</button>
-  <small id="genstat_numrows_text">Showing <span id="genstat_numrows">{{ report.general_stats_html['rows'].keys() | length }}</span> rows.</small>
-</h2>
-<div id="general_stats_table_container">
-  <div class="table-responsive">
-    <table id="general_stats_table" class="table table-condensed mqc_table">
-      <thead>
-        <tr>
-          <th class="rowheader">Sample Name</th>
-          {%- for k, h in report.general_stats_html['headers'].items() %}
-            {{ h }}
-          {%- endfor -%}
-        </tr>
-      </thead>
-      <tbody>
-        {% for sn, r in report.general_stats_html['rows']|dictsort %}
-        <tr>
-          <th class="rowheader" data-original-sn="{{ sn }}">{{ sn }}</th>
-          {%- for k, h in report.general_stats_html['headers'].items() %}
-            {{ r[k] if r[k] else '<td class="'+k+'"></td>' }}
-          {%- endfor -%}
-        </tr>
-        {%- endfor %}
-      </tbody>
-    </table>
-  </div>
-</div>
-</div>
-
-
-
-
-<!-- General Stats Columns Modal -->
-<div class="modal fade" id="general_stats_colsort_modal" tabindex="-1">
-  <div class="modal-dialog modal-lg">
-    <div class="modal-content">
-      <div class="modal-header">
-        <button type="button" class="close" data-dismiss="modal" aria-label="Close"><span aria-hidden="true">&times;</span></button>
-        <h4 class="modal-title">General Statistics: Columns</h4>
-      </div>
-      <div class="modal-body">
-        <p>Uncheck the tick box to hide columns. Click and drag the handle on the left to change order.</p>
-        <table class="table mqc_table" id="general_stats_colsort_table">
-          <thead>
-            <tr>
-              <th class="sorthandle" style="text-align:center;">Sort</th>
-              <th style="text-align:center;">Visible</th>
-              <th>Module</th>
-              <th>Column</th>
-              <th>Description</th>
-              <th>Scale</th>
-            </tr>
-          </thead>
-          <tbody>
-            {%- for m in report.general_stats -%}{%- for k in report.general_stats[m]['headers'].keys() -%}{%- if report.general_stats[m]['headers'][k]['rid'] in report.general_stats_html['headers'] %}
-              <tr class="{{ report.general_stats[m]['headers'][k]['rid'] }}" style="background-color: rgba({{report.general_stats[m]['headers'][k]['modcol'] }}, 0.15);">
-                <td class="sorthandle ui-sortable-handle">||</span></td>
-                <td style="text-align:center;"><input class="general_stats_col_visible" type="checkbox" checked="checked" value="{{ report.general_stats[m]['headers'][k]['rid'] }}"></td>
-                <td>{{ m }}</td>
-                <td>{{ report.general_stats[m]['headers'][k]['title'] }}</td>
-                <td>{{ report.general_stats[m]['headers'][k]['description'] }}</td>
-                <td>{{ report.general_stats[m]['headers'][k]['shared_key'] }}</td>
-              </tr>
-            {%- endif -%}{%- endfor -%}{%- endfor -%}
-          </tbody>
-        </table>
-      </div>
-      <div class="modal-footer">
-        <button type="button" class="btn btn-default" data-dismiss="modal">Close</button>
-      </div>
-    </div>
-  </div>
-</div>
-
-{% endif %}
